Refactor functions into their own modules
This makes it much easier to mock things in the future, and also
structures the code better.

This commit is best viewed with `--color-moved`.
